+++ src/MathObjects/vector.cpp
@@ -1,7 +1,7 @@
 #include "vector.h"
 
 
-Vector::Vector(float x, float y, float z)
+Vector::Vector(double x, double y, double z)
 {
     this->x = x;
     this->y = y;
